OAuth2/requirements.txt
alembic
annotated-types
anyio
bcrypt
certifi
click
colorama
dnspython
email_validator
fastapi
fastapi-cli
greenlet
h11
httpcore
httptools
httpx
idna
iniconfig
itsdangerous
Jinja2
Mako
markdown-it-py
MarkupSafe
mdurl
orjson
packaging
passlib[argon2, bcrypt]
pwdlib[argon2, bcrypt]
pluggy
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
PyJWT
pytest
python-dotenv
python-multipart
PyYAML
rich
shellingham
sniffio
SQLAlchemy
starlette
typer
typing_extensions
tzdata
ujson
uvicorn
watchfiles
websockets
